Refactor Detector for handling generic type and pre-added more data for SeekTactician to track pheromones

+++ Assets/Library/StateMachines/Gland/NonSerializedInfo.cs
@@ -11,7 +11,7 @@ namespace Assets.Library.StateMachines.Gland
         public UnityEvent<Storage> ContactWithStorageHappened { get; set; }
         public UnityEvent<Collectable> ContactWithNonEmptyCollectableHappened { get; set; }
         public BoundedUint Load { get; set; }
-        public UnityAction<GameObject, Identifier, float> InitPheromone { get; set; }
+        public UnityAction<GameObject, Identifier, uint> InitPheromone { get; set; }
         public Identifier Identifier { get; set; }
 
         public void CheckValidity()
